pyproject.toml
[project]
name = "dvm-mesura"
version = "0.1.0"
description = "Home automation monitoring suite for energy (HomeWizard), weather (OpenWeatherMap), and heating (Honeywell Evohome)"
readme = "README.md"
authors = [
    { name = "divyavanmahajan", email = "divyavanmahajan@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "evohome-async>=0.4.15",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tabulate>=0.9.0",
]

[project.scripts]
mesura-energy = "dvm_mesura.energymeter:main"
mesura-weather = "dvm_mesura.openweathermap:main"
mesura-evohome = "dvm_mesura.evohome:main_cli"
mesura-all = "dvm_mesura.main:main"

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "responses>=0.26.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-sugar>=1.1.1",
]
